Add ArcadeBadge and ArcadeActivity components; update import paths in IndexPopup and ArcadeCard
